(maint) Add support for passing an image to test

+++ spec/support/context/with_docker_image.rb
@@ -1,12 +1,19 @@
 require 'json'
 shared_context 'with a docker image' do
   before(:all) do
-    @image = "test/#{File.basename(CURRENT_DIRECTORY)}:#{Random.rand(1000)}"
-    %x(docker image build --tag #{@image} #{CURRENT_DIRECTORY})
+    if ENV['PUPPET_TEST_DOCKER_IMAGE'] && !ENV['PUPPET_TEST_DOCKER_IMAGE'].empty?
+      @image=ENV['PUPPET_TEST_DOCKER_IMAGE']
+    else
+      @image = "test/#{File.basename(CURRENT_DIRECTORY)}:#{Random.rand(1000)}"
+      %x(docker image build --tag #{@image} #{CURRENT_DIRECTORY})
+    end
+    puts "Running tests on #{@image}..."
     @image_json = JSON.parse(%x(docker inspect #{@image}))
   end
 
   after(:all) do
-    %x(docker image rm --force #{@image})
+    if !ENV['PUPPET_TEST_DOCKER_IMAGE'] || ENV['PUPPET_TEST_DOCKER_IMAGE'].empty?
+      %x(docker image rm --force #{@image})
+    end
   end
 end
